+++ rdsai-20250507/alibabacloud_rdsai20250507/models/__init__.py
@@ -85,6 +85,10 @@ from ._modify_instance_storage_config_request import ModifyInstanceStorageConfig
 from ._modify_instance_storage_config_shrink_request import ModifyInstanceStorageConfigShrinkRequest
 from ._modify_instance_storage_config_response_body import ModifyInstanceStorageConfigResponseBody
 from ._modify_instance_storage_config_response import ModifyInstanceStorageConfigResponse
+from ._modify_instances_sslrequest import ModifyInstancesSSLRequest
+from ._modify_instances_sslshrink_request import ModifyInstancesSSLShrinkRequest
+from ._modify_instances_sslresponse_body import ModifyInstancesSSLResponseBody
+from ._modify_instances_sslresponse import ModifyInstancesSSLResponse
 from ._modify_messages_feedbacks_request import ModifyMessagesFeedbacksRequest
 from ._modify_messages_feedbacks_response_body import ModifyMessagesFeedbacksResponseBody
 from ._modify_messages_feedbacks_response import ModifyMessagesFeedbacksResponse
@@ -116,6 +120,7 @@ from ._describe_instance_ip_whitelist_response_body import DescribeInstanceIpWhi
 from ._describe_instance_ragconfig_response_body import DescribeInstanceRAGConfigResponseBodyConfigList
 from ._describe_instance_storage_config_response_body import DescribeInstanceStorageConfigResponseBodyConfigList
 from ._get_conversations_response_body import GetConversationsResponseBodyData
+from ._get_messages_response_body import GetMessagesResponseBodyDataEvents
 from ._get_messages_response_body import GetMessagesResponseBodyData
 from ._list_custom_agent_response_body import ListCustomAgentResponseBodyData
 from ._list_custom_agent_tools_response_body import ListCustomAgentToolsResponseBodyData
@@ -207,6 +212,10 @@ __all__ = [
     ModifyInstanceStorageConfigShrinkRequest,
     ModifyInstanceStorageConfigResponseBody,
     ModifyInstanceStorageConfigResponse,
+    ModifyInstancesSSLRequest,
+    ModifyInstancesSSLShrinkRequest,
+    ModifyInstancesSSLResponseBody,
+    ModifyInstancesSSLResponse,
     ModifyMessagesFeedbacksRequest,
     ModifyMessagesFeedbacksResponseBody,
     ModifyMessagesFeedbacksResponse,
@@ -238,6 +247,7 @@ __all__ = [
     DescribeInstanceRAGConfigResponseBodyConfigList,
     DescribeInstanceStorageConfigResponseBodyConfigList,
     GetConversationsResponseBodyData,
+    GetMessagesResponseBodyDataEvents,
     GetMessagesResponseBodyData,
     ListCustomAgentResponseBodyData,
     ListCustomAgentToolsResponseBodyData,